+++ core/golang_internal_testsuite/static_files_consts/static_file_consts.go
@@ -5,23 +5,12 @@
 
 package static_files_consts
 
-import (
-	"github.com/kurtosis-tech/kurtosis-client/golang/lib/services"
-	"path"
-)
-
 const (
 	// Directory where static files live inside the testsuite container
-	staticFilesDirpathOnTestsuiteContainer = "/static-files"
-
-	TestStaticFile1ID      services.StaticFileID = "test-static-file1"
-	TestStaticFile2ID	   services.StaticFileID = "test-static-file2"
+	StaticFilesDirpathOnTestsuiteContainer = "/static-files"
 
 	testStaticFile1Filename = "test-static-file1.txt"
 	testStaticFile2Filename = "test-static-file2.txt"
 )
 
-var StaticFileFilepaths = map[services.StaticFileID]string{
-	TestStaticFile1ID: path.Join(staticFilesDirpathOnTestsuiteContainer, testStaticFile1Filename),
-	TestStaticFile2ID: path.Join(staticFilesDirpathOnTestsuiteContainer, testStaticFile2Filename),
-}
+var StaticFilesNames = []string{testStaticFile1Filename, testStaticFile2Filename}